Restructure and fix bugs.
