+++ .gitignore
@@ -70,3 +70,7 @@ bower_components/
 .grunt/
 src/vendor/
 dist/
+
+# Build #
+#########
+build/
