+++ cli/gocd/gocd_structs.go
