+++ cli/report/pongoReport.html
@@ -1,0 +1,13 @@
+<html><head></head><body><h2>Test type: pongo2</h2>
+
+ 
+   Some: simple 
+   Thing: simple
+
+
+ 
+   Some: simple 
+   Thing: simple
+
+
+</body></html>
