+++ cli/report/report_test.go
@@ -6,8 +6,8 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 
-	"github.com/cfpb/rhobot/config"
-	"github.com/cfpb/rhobot/database"
+	"github.com/cfpb/rhobot/cli/config"
+	"github.com/cfpb/rhobot/cli/database"
 )
 
 var conf *config.Config
