+++ cmd/rhobot/main_test.go
@@ -7,10 +7,10 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 
-	"github.com/cfpb/rhobot/config"
-	"github.com/cfpb/rhobot/database"
-	"github.com/cfpb/rhobot/healthcheck"
-	"github.com/cfpb/rhobot/report"
+	"github.com/cfpb/rhobot/cli/config"
+	"github.com/cfpb/rhobot/cli/database"
+	"github.com/cfpb/rhobot/cli/healthcheck"
+	"github.com/cfpb/rhobot/cli/report"
 )
 
 var conf *config.Config
