+++ cmd/rhobot/util.go
@@ -8,11 +8,11 @@ import (
 	"time"
 
 	log "github.com/Sirupsen/logrus"
-	"github.com/cfpb/rhobot/config"
-	"github.com/cfpb/rhobot/database"
-	"github.com/cfpb/rhobot/gocd"
-	"github.com/cfpb/rhobot/healthcheck"
-	"github.com/cfpb/rhobot/report"
+	"github.com/cfpb/rhobot/cli/config"
+	"github.com/cfpb/rhobot/cli/database"
+	"github.com/cfpb/rhobot/cli/gocd"
+	"github.com/cfpb/rhobot/cli/healthcheck"
+	"github.com/cfpb/rhobot/cli/report"
 	"github.com/davecgh/go-spew/spew"
 	"github.com/urfave/cli"
 )
