+++ cmd/rhobot/version.go
@@ -1,4 +1,4 @@
 package main
 
 // Version of Rhobot
-const Version = `0.4.1`
+const Version = `0.4.2`
