+++ go.mod
@@ -1,0 +1,24 @@
+module github.com/cfpb/rhobot
+
+go 1.15
+
+require (
+	github.com/PuerkitoBio/goquery v1.1.1-0.20170816133831-ce645ea5e67a
+	github.com/Sirupsen/logrus v1.0.4-0.20170822132746-89742aefa4b2
+	github.com/andybalholm/cascadia v0.0.0-20161224141413-349dd0209470
+	github.com/davecgh/go-spew v1.1.1-0.20170711183451-adab96458c51
+	github.com/flosch/pongo2 v0.0.0-20170704123420-58f1f3387f7c
+	github.com/gorilla/css v1.0.0
+	github.com/juju/errors v0.0.0-20170703010042-c7d06af17c68
+	github.com/lib/pq v0.0.0-20170810061220-e42267488fe3
+	github.com/pmezard/go-difflib v1.0.0
+	github.com/stretchr/testify v1.1.5-0.20170809224252-890a5c3458b4
+	github.com/urfave/cli v1.20.1-0.20170813145949-f017f86fccc5
+	github.com/vanng822/css v0.0.0-20161108042044-00dd67448c5d
+	github.com/vanng822/go-premailer v0.0.0-20170721182134-5de9db06a79a
+	golang.org/x/crypto v0.0.0-20170825220121-81e90905daef
+	golang.org/x/net v0.0.0-20170824174428-57efc9c3d9f9
+	golang.org/x/sys v0.0.0-20170825194620-2d6f6f883a06
+	gopkg.in/gomail.v2 v2.0.0-20160411212932-81ebce5c23df
+	gopkg.in/yaml.v2 v2.0.0-20170812160011-eb3733d160e7
+)
